@@ -1,26 +1,26 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Azure.Extensions.AspNetCore.DataProtection.Blobs" Version="1.3.4" />
-		<PackageReference Include="Azure.Extensions.AspNetCore.DataProtection.Keys" Version="1.2.3" />
-		<PackageReference Include="Azure.Identity" Version="1.12.0" />
-		<PackageReference Include="Blazor.Serialization" Version="8.0.0" />
-		<PackageReference Include="Blazor.LocalStorage" Version="8.0.0" />
-		<PackageReference Include="Bogus" Version="35.6.0" />
-		<PackageReference Include="JetBrains.Annotations" Version="2024.2.0" />
+		<PackageReference Include="Azure.Extensions.AspNetCore.DataProtection.Blobs" Version="1.4.0" />
+		<PackageReference Include="Azure.Extensions.AspNetCore.DataProtection.Keys" Version="1.3.0" />
+		<PackageReference Include="Azure.Identity" Version="1.13.1" />
+		<PackageReference Include="Blazor.Serialization" Version="9.0.0" />
+		<PackageReference Include="Blazor.LocalStorage" Version="9.0.0" />
+		<PackageReference Include="Bogus" Version="35.6.1" />
+		<PackageReference Include="JetBrains.Annotations" Version="2024.3.0" />
 		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.22.0" />
-		<PackageReference Include="Microsoft.Azure.SignalR" Version="1.26.0" />
-		<PackageReference Include="Microsoft.Extensions.Azure" Version="1.7.4" />
-		<PackageReference Include="Microsoft.Orleans.Clustering.AzureStorage" Version="8.2.0" />
-		<PackageReference Include="Microsoft.Orleans.Client" Version="8.2.0" />
+		<PackageReference Include="Microsoft.Azure.SignalR" Version="1.29.0" />
+		<PackageReference Include="Microsoft.Extensions.Azure" Version="1.9.0" />
+		<PackageReference Include="Microsoft.Orleans.Clustering.AzureStorage" Version="9.0.1" />
+		<PackageReference Include="Microsoft.Orleans.Client" Version="9.0.1" />
 		<PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.21.0" />
-		<PackageReference Include="MudBlazor" Version="7.4.0" />
+		<PackageReference Include="MudBlazor" Version="7.15.0" />
 	</ItemGroup>
 
 	<ItemGroup>
